Move rate modules to library
This allows us to dumb down the binaries and move the tests into the
appropriate place i.e. next to where the function-under-test is defined.
